% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fineboost_updates.R
\name{calc_step}
\alias{calc_step}
\title{Get optimal step sizes for boosting updates}
\usage{
calc_step(xx, yy, curr_step)
}
\arguments{
\item{xx}{A vector of length N (arbitrary).}

\item{yy}{A second vector of length N.}

\item{curr_step}{The stepsize at the start of the update.}
}
\value{
The optimal step size for the next update of the coefficients.
}
\description{
Get optimal step sizes for boosting updates
}
