% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fineboost_utils.R
\name{calc_tau}
\alias{calc_tau}
\title{Exploring results for different values of kernel smoothing parameter}
\usage{
calc_tau(X, Y, maxsize = 10)
}
\arguments{
\item{X}{The design matrix X (N times P) with samples/individuals along the rows
and putatively correlated ordered features (SNPs) along the columns.}

\item{Y}{The response vector of length N}

\item{maxsize}{the maximum number of non-zero causal loci based on L0Learn.}
}
\value{
The objective function calculation for different choices of tau.
}
\description{
Exploring results for different values of kernel smoothing parameter
}
