% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_X_attributes.R
\name{compute_colnorm}
\alias{compute_colnorm}
\title{computes column L2 norm for any type of matrix}
\usage{
compute_colnorm(X)
}
\arguments{
\item{X}{an n by p matrix of any type, e.g. sparse, dense}
}
\value{
a p vector of column L-2 norms
}
\description{
computes column L2 norm for any type of matrix
}
