% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fineboost_utils.R
\name{fineboost_get_ccg}
\alias{fineboost_get_ccg}
\title{Compute inclusion confidence grade (CCG) for all variables}
\usage{
fineboost_get_ccg(ff, use_csets = TRUE)
}
\arguments{
\item{ff}{an object of class fineboos, generated from `fineboostR::fineboost_smooth()`}

\item{use_csets}{Whether to use the confidence sets when computing the ICG. Defaults to TRUE.}
}
\value{
a vector inclusion confidence grade for all variables.
}
\description{
Compute inclusion confidence grade (CCG) for all variables
}
