% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fineboost_get_coef.R
\name{fineboost_get_coef}
\alias{fineboost_get_coef}
\title{Generate coefficient estimates using the FINEBoost model}
\usage{
fineboost_get_coef(X, y, ff)
}
\arguments{
\item{X}{The design matrix X (N times P) with samples/individuals along the rows
and putatively correlated ordered features (SNPs) along the columns.}

\item{y}{The response vector of length N}

\item{ff}{An object of class fineboost, containing the fitted model with ccg and confidence sets information.}
}
\value{
Returns a P dimensional vector b in \eqn{y = Xb + e} based on the FINEBoost model fit
}
\description{
Generate coefficient estimates using the FINEBoost model
}
