% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fineboost_updates.R
\name{update_kernel_weights}
\alias{update_kernel_weights}
\title{One instance of the kernel-smoothed FS-boost weights update}
\usage{
update_kernel_weights(X, y, LD, tau)
}
\arguments{
\item{X}{The design matrix X (N times P) with samples/individuals along the rows
and putatively correlated ordered features (SNPs) along the columns.}

\item{y}{The response vector of length N}

\item{LD}{The LD matrix for all P variables in the X matrix.}

\item{tau}{The kernel smoothing bandwidth estimate.}
}
\value{
An object with the following items
  \item{weights}{A new vector of weights of length P based on kernel smoothing around the top feature.}
  \item{objective}{The current value of the objective to be minimized.}
  \item{corvals}{The current values of correlations of the P features with the current residual.}
}
\description{
One instance of the kernel-smoothed FS-boost weights update
}
