% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fineboost_utils.R
\name{fineboost_get_csets}
\alias{fineboost_get_csets}
\title{Get confidence sets for local signal clusters from FineBoost output}
\usage{
fineboost_get_csets(ff, coverage = 0.95, clus_thresh = 0.1,
  nmf_try = 10)
}
\arguments{
\item{ff}{an object of class fineboos, generated from `fineboostR::fineboost_smooth()`}

\item{coverage}{A number between 0 and 1 (close to 1) specifying the coverage of the estimated signal clusters.
Default set to 0.95.}

\item{clus_thresh}{A number between 0 and 1 (close to 0) that is used to filter out local signal clusters with
depleted number of boosting iterations and high level of uniformity of signal. Default is
set to 0.1.}

\item{nmf_try}{Number of initializations of the Non-negative matrix factorization on the trajectory of weights
tested. Default is set to 10.}

\item{LD}{the LD P times P matrix, where P is the number of features, equal to ff$P,}

\item{min_abs_corr}{Minimum of absolute value of correlation allowed in a credible set. The default, 0.5,
corresponds to squared correlation of 0.25, which is a commonly used threshold for
genotype data in genetics studies.}
}
\value{
An object with the following items
  \item{evidence}{A vector of length `Lmax` giving the evidence strength for each of signal clusters.}
  \item{unfiltered_sets}{A list of length `Lmax` denoting the confidence sets that contain causal signal, where the
                        strength of confidence is provided by the coverage.}
  \item{cluster_signal}{The cluster signal for the `unfiltered sets`.}
  \item{filtered_sets}{The threshold used to filter away cluster signals of the evidence for that cluster falls below this
                       threshold. Default is set at 0.1.}
  \item{csets_index}{A vector of length P that labels each variable by the confidence set in the `filtered_sets` it
                     belongs to and 0 otherwise}
}
\description{
Get confidence sets for local signal clusters from FineBoost output
}
